+++ syncforge/src/commonMain/kotlin/dev/syncforge/entity/EntityStoreSyncHandler.kt
@@ -1,13 +1,10 @@
 package dev.syncforge.entity
 
-import dev.syncforge.api.ExperimentalSyncForgeApi
-
 /**
  * [TypedEntitySyncHandler] backed by an [EntityStore] instead of DAO-shaped insert/update/delete.
  *
  * Subclasses supply JSON mapping and [withSyncState]; CRUD delegates to [store].
  */
-@ExperimentalSyncForgeApi
 abstract class EntityStoreSyncHandler<T : SyncedEntity>(
     protected val store: EntityStore<T>,
 ) : TypedEntitySyncHandler<T>() {